* Falta terminar a normalização dos pesos de arestas em Graph para poder iniciar as alterações na similaridade do SCAN.
git-svn-id: http://home.speed.dcc.ufmg.br:8080/svn/public/helio/scan@4862 c489cccc-4d05-0410-97be-cfb9d72d6f98

+++ scan.hpp
@@ -14,19 +14,19 @@ class Scan{
           uint num_clusters;
           ScanGraph g;
           // Methods
-          bool isCore(uint node, const float epsilon, const int mi);
-          float similarity(uint node1, uint node2);
+          bool isCore(uint node, const double epsilon, const int mi);
+          double similarity(uint node1, uint node2);
           int getNewClusterID();
           std::set<Edge> dirReach(uint v, 
-                    const float epsilon, const int mi);
+                    const double epsilon, const int mi);
           void addToCluster(uint node, uint clust);
           std::set<uint> getNeighborClusters(uint node);
           void addHub(uint node, std::set<uint> clusts);
           std::set<Edge> neighborhood(uint node, 
-                    const float epsilon);
+                    const double epsilon);
           void buildAssortativityMatrix(float** e);
      public:
-          void run(const float epsilon, const int mi);
+          void run(const double epsilon, const int mi);
           Scan();
           Scan(std::string filename);
           void loadGraph(std::string filename);
